+++ crates/mako/src/compiler.rs
@@ -236,7 +236,7 @@ impl Compiler {
             Arc::new(plugins::wasm_runtime::WasmRuntimePlugin {}),
             Arc::new(plugins::async_runtime::AsyncRuntimePlugin {}),
             Arc::new(plugins::emotion::EmotionPlugin {}),
-            Arc::new(plugins::tree_shake::FarmTreeShake {}),
+            Arc::new(plugins::tree_shaking::FarmTreeShake {}),
         ];
         plugins.extend(builtin_plugins);
