+++ crates/mako/src/plugins/tree_shaking.rs
