+++ crates/mako/src/plugins/tree_shaking/module_side_effects_flag.rs
