+++ crates/mako/src/plugins/tree_shaking/shake/module_concatenate/concatenated_transformer/utils.rs
