+++ crates/mako/src/plugins/tree_shaking/shake/skip_module.rs
@@ -14,9 +14,9 @@ use swc_core::quote;
 use crate::compiler::Context;
 use crate::module::{Dependency, ImportType, ModuleId, NamedExportType, ResolveType};
 use crate::module_graph::ModuleGraph;
-use crate::plugins::tree_shake::module::{is_ident_sym_equal, TreeShakeModule};
-use crate::plugins::tree_shake::shake::strip_context;
-use crate::plugins::tree_shake::statement_graph::{
+use crate::plugins::tree_shaking::module::{is_ident_sym_equal, TreeShakeModule};
+use crate::plugins::tree_shaking::shake::strip_context;
+use crate::plugins::tree_shaking::statement_graph::{
     ExportSpecifierInfo, ImportSpecifierInfo, StatementId,
 };
